+++ tests/ImageSharp.Tests/Processing/Effects/ContrastTest.cs
@@ -1,0 +1,53 @@
+﻿// <copyright file="ContrastTest.cs" company="James Jackson-South">
+// Copyright (c) James Jackson-South and contributors.
+// Licensed under the Apache License, Version 2.0.
+// </copyright>
+
+namespace ImageSharp.Tests.Processing.Effects
+{
+    using ImageSharp.PixelFormats;
+
+    using Xunit;
+
+    public class ContrastTest : FileTestBase
+    {
+        public static readonly TheoryData<int> ContrastValues
+        = new TheoryData<int>
+        {
+            50,
+           -50
+        };
+
+        [Theory]
+        [WithFileCollection(nameof(DefaultFiles), nameof(ContrastValues), StandardPixelType)]
+        public void ImageShouldApplyContrastFilter<TPixel>(TestImageProvider<TPixel> provider, int value)
+            where TPixel : struct, IPixel<TPixel>
+        {
+            using (Image<TPixel> image = provider.GetImage())
+            {
+                image.Contrast(value)
+                    .DebugSave(provider, value, Extensions.Bmp);
+            }
+        }
+
+        [Theory]
+        [WithFileCollection(nameof(DefaultFiles), nameof(ContrastValues), StandardPixelType)]
+        public void ImageShouldApplyContrastFilterInBox<TPixel>(TestImageProvider<TPixel> provider, int value)
+            where TPixel : struct, IPixel<TPixel>
+        {
+            using (Image<TPixel> source = provider.GetImage())
+            using (var image = new Image<TPixel>(source))
+            {
+                var bounds = new Rectangle(10, 10, image.Width / 2, image.Height / 2);
+
+                image.Contrast(value, bounds)
+                    .DebugSave(provider, value, Extensions.Bmp);
+
+                // Draw identical shapes over the bounded and compare to ensure changes are constrained.
+                image.Fill(NamedColors<TPixel>.HotPink, bounds);
+                source.Fill(NamedColors<TPixel>.HotPink, bounds);
+                ImageComparer.CheckSimilarity(image, source);
+            }
+        }
+    }
+}
