--- tests/ImageSharp.Tests/Processors/Filters/ContrastTest.cs
@@ -1,56 +1,0 @@
-﻿// <copyright file="ContrastTest.cs" company="James Jackson-South">
-// Copyright (c) James Jackson-South and contributors.
-// Licensed under the Apache License, Version 2.0.
-// </copyright>
-
-namespace ImageSharp.Tests
-{
-    using System.IO;
-
-    using ImageSharp.PixelFormats;
-
-    using Xunit;
-
-    public class ContrastTest : FileTestBase
-    {
-        public static readonly TheoryData<int> ContrastValues
-        = new TheoryData<int>
-        {
-            50 ,
-           -50 ,
-        };
-
-        [Theory]
-        [MemberData(nameof(ContrastValues))]
-        public void ImageShouldApplyContrastFilter(int value)
-        {
-            string path = this.CreateOutputDirectory("Contrast");
-
-            foreach (TestFile file in Files)
-            {
-                using (Image<Rgba32> image = file.CreateImage())
-                using (FileStream output = File.OpenWrite($"{path}/{file.FileName}"))
-                {
-                    image.Contrast(value).Save(output);
-                }
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(ContrastValues))]
-        public void ImageShouldApplyContrastFilterInBox(int value)
-        {
-            string path = this.CreateOutputDirectory("Contrast");
-
-            foreach (TestFile file in Files)
-            {
-                string filename = file.GetFileName(value + "-InBox");
-                using (Image<Rgba32> image = file.CreateImage())
-                using (FileStream output = File.OpenWrite($"{path}/{filename}"))
-                {
-                    image.Contrast(value, new Rectangle(10, 10, image.Width / 2, image.Height / 2)).Save(output);
-                }
-            }
-        }
-    }
-}
